Move Nav directory to client/src/layouts
Also moved tests with it

Updated imports as needed

Moved LayoutWrapper and test to the same directory
